+++ CHANGELOG.md
@@ -9,5 +9,7 @@
 * Clients\CommandRunner added
 * Exceptions\E4xx_CommandFailed added
 * Exceptions\E4xx_CommandSucceeded added
+* Requirements\RequireCommandFailed added
+* Requirements\RequireCommandSucceeded added
 * Values\CommandResult added
 * ValueBuilders\BuildEscapedCommand added